Add category data migration
