+++ .gitignore
@@ -2,3 +2,4 @@ venv
 simple.db
 __pycache__
 .DS_Store
+.vscode
