+++ scrapy.cfg
@@ -4,8 +4,8 @@
 # https://scrapyd.readthedocs.io/en/latest/deploy.html
 
 [settings]
-default = vintourage.crawler.settings
+default = crawler.settings
 
 [deploy]
 #url = http://localhost:6800/
-project = vintourage.crawler
+project = crawler
