+++ tests/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+from vintourage import app
+
+@pytest.fixture
+def app():
+    app = create_app()
+    return app
